tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.pattern.modify.canvas1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.pattern.modify.canvas1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.pattern.modify.canvas1</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var canvas2 = new OffscreenCanvas(100, 50);
  var ctx2 = canvas2.getContext('2d');
  ctx2.fillStyle = '#0f0';
  ctx2.fillRect(0, 0, 100, 50);

  var pattern = ctx.createPattern(canvas2, 'no-repeat');

  ctx2.fillStyle = '#f00';
  ctx2.fillRect(0, 0, 100, 50);

  ctx.fillStyle = pattern;
  ctx.fillRect(0, 0, 100, 50);

  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 98,1, 0,255,0,255);
  _assertPixel(canvas, 1,48, 0,255,0,255);
  _assertPixel(canvas, 98,48, 0,255,0,255);
  t.done();

});
</script>
